requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
requests==2.31.0
httpx==0.25.0
python-dotenv==1.0.0
python-multipart==0.0.6
tensorflow==2.13.0
pillow==10.0.1
numpy==1.24.3
pydantic==2.3.0